Fix error in RosterRemoveUser's use of group names
This fix isn't especially elegant, but I don't want to dig any more deeply into this code than I have to right now :)

The groups that are stored in the user object seem to be strings, rather than actual group objects that have a name attribute. If RosterRemoveUser isn't able to determine the name of the group, then when it tries to remove the user's entry from the buddy list, it will fail, and there will be phantom entries after the roster item has been removed.