Updated file paths to work with nginx and vine-web
JWChat assumes that all it's files are served out of the same directory as the main window is launched from, which is no good if I want to include them all as static files that are part of a larger nginx/gunicorn project. I can tell nginx to serve the static files out of /static/jwchat/, and then I need to update *all* of the hard-coded paths so that it can find everything. This was a pain, but TextMate's regex find-replace helped a lot. Hopefully I didn't miss any!